Change type of database year to NSInteger.  Be more careful about initialization and parsing.

+++ TLMDatabase.h
@@ -38,7 +38,7 @@
 
 #import <Cocoa/Cocoa.h>
 
-typedef int32_t TLMDatabaseYear;
+typedef NSInteger TLMDatabaseYear;
 
 // returned as the year in case of an error
 extern const TLMDatabaseYear TLMDatabaseUnknownYear;
